requirements.txt
celery==4.4.7
Django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.1.0
flower==0.9.7
gunicorn==21.2.0
isort==5.12.0
python-dotenv==1.0.0
redis==3.5.3
requests==2.31.0
psycopg2-binary==2.9.7
